codex: fix CI failure on PR #16201
